SPRNET-1273 - Remove unnecessary cast to (int) in Spring.Data.Core.AdoTemplate ExecuteScalar with ICommandSetter

+++ src/Spring/Spring.Data/Data/Core/AdoTemplate.cs
@@ -532,7 +532,7 @@ namespace Spring.Data.Core
             }
             #endregion
 
-            return (int)Execute(new ExecuteScalarCallbackWithCommandSetter(cmdType, cmdText, commandSetter));
+            return Execute(new ExecuteScalarCallbackWithCommandSetter(cmdType, cmdText, commandSetter));
 
         }
